:fire: chore(auth): consolidate and refactor email verification controllers
Consolidate the email verification logic from three separate controllers
into a single `EmailVerificationController` for better maintainability
and structure.

- Rename `VerifyEmailController` to `EmailVerificationController`.
- Merge the logic from `EmailVerificationNotificationController` and
`EmailVerificationPromptController` into the new
`EmailVerificationController`.
- Update routing in `routes/auth.php` to use the new
`EmailVerificationController` methods (`show`, `store`, `verify`).
- Introduce `.php-cs-fixer.dist.php` configuration file to enforce code
style standards.
- Minor cleanup in `AuthenticatedSessionController` by removing an
unused import.
